update truffle config.js
the truffle config.js file determines how the contract will be migrated to the network. In this case, we need to specify a specific solidity compiler just so that we don't run into any wierd issues when compiling older contracts, and we need to instantiate a new Wallet with HDWalletProvider given our user-specific mnemonic and RPC endpoint. The mnemonic is given when we create a metamask wallet and the RPC enpoint is taken from the ethereum on azure output log from our master node.