add packages and change package name
a few packages have been added and updated here including a local version of truffle, truffle-hdwallet-provider, and an update to web3 v1.0.0-beta.34 which was neccessary due to indeterminable issues with launching the client. dotenv was also added in order to obfuscate environment variables.

The package name was changed from rate my prof to rate my hack